Uploaded bubblesort code
getList: User inputs list. Program checks list to see if there are 2-8 integers of values 0-9, returns values if they are okay.
getListAmt: Program checks list and determines which integers are input by the user(the rest of the array is filled with -1.
bubblesort: Program sorts the values of the list, printing each time the program goes through.